mati-test/requirements.txt
google-transit-gtfs-realtime-bindings>=1.0.0
requests>=2.31.0
flask>=3.0.0
protobuf>=4.0.0

